+++ .vscode/cspell.json
@@ -63,6 +63,7 @@
         "notequals",
         "notna",
         "numpy",
+        "ONHOLD",
         "pformat",
         "pickel",
         "pylint",
